fix(firmware): 修复后台刷新完成状态机
- 拆分 idle/done publishing 与 queued 状态避免抢占消费丢事件
- 后台刷新 done 携带 generation 并由 App 校验当前生命周期
- 补可重入消费、释放 owner、发布争用和 generation 回归测试

+++ firmware/main/events/ui_event.h
@@ -131,7 +131,7 @@ struct UiEvent {
             uint32_t token;
         } xiaozhi_channel;
         struct {
-            uint32_t generation;
+            uint64_t generation;
         } bg_refresh;
         U() : group{} {
         }
